# QueR.Application/swagger-backoffice.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "QueR Backoffice API",
    "description": "QueR Backoffice API documentation for development purposes.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/backoffice/Queues/{queueId}": {
      "put": {
        "tags": [
          "Queues"
        ],
        "operationId": "Queues_UpdateQueue",
        "parameters": [
          {
            "name": "queueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueModel"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Queues"
        ],
        "operationId": "Queues_GetDetailsOfQueue",
        "parameters": [
          {
            "name": "queueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/backoffice/QueueTypes": {
      "post": {
        "tags": [
          "QueueTypes"
        ],
        "operationId": "QueueTypes_CreateQueueType",
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueTypeModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTypeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/backoffice/QueueTypes/{queueTypeId}": {
      "put": {
        "tags": [
          "QueueTypes"
        ],
        "operationId": "QueueTypes_UpdateQueueType",
        "parameters": [
          {
            "name": "queueTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueTypeModel"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "QueueTypeDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "numOfQueues": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "QueueDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "queueType": {
            "type": "string",
            "nullable": true
          },
          "queueTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "nextNumber": {
            "type": "integer",
            "format": "int32"
          },
          "worksite": {
            "type": "string",
            "nullable": true
          },
          "prefix": {
            "type": "string",
            "nullable": true
          },
          "step": {
            "type": "integer",
            "format": "int32"
          },
          "worksiteId": {
            "type": "integer",
            "format": "int32"
          },
          "numOfAssignedEmployees": {
            "type": "integer",
            "format": "int32"
          },
          "numOfActiveTickets": {
            "type": "integer",
            "format": "int32"
          },
          "maxActiveTicketsPerUser": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "QueueModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "typeId": {
            "type": "integer",
            "format": "int32"
          },
          "nextNumber": {
            "type": "integer",
            "format": "int32"
          },
          "prefix": {
            "type": "string",
            "nullable": true
          },
          "step": {
            "type": "integer",
            "format": "int32"
          },
          "maxActiveTicketsPerUser": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "QueueTypeModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}